+++ game/traildrawer.py
@@ -1,7 +1,7 @@
 import pbge
 
 def draw_trail( sprite
-              , trailmarker, zero
+              , trailmarker, zero, endcursor
               , scene, mover
               , mp_remaining
               , path
@@ -9,7 +9,7 @@ def draw_trail( sprite
     ''' Common code to draw a trail, showing where one action point
     is completely spent.
     '''
-    aps = 0 # Action points so far
+    aps = 1 # Action points so far
     for i in range(len(path) - 1):
         p = path[i]
         nextp = path[i + 1]
@@ -23,3 +23,4 @@ def draw_trail( sprite
         mp_remaining -= cost
         if i != 0:
             pbge.my_state.view.overlays[p] = (sprite, marker)
+    pbge.my_state.view.overlays[path[-1]] = (sprite, endcursor)