+++ history.md
@@ -1,3 +1,4 @@
+* Added combat action clock
 * Reworked the combat movement UI
 * Fixed Ammo Explosion display bug
 * Extreme speeds will now be slightly reduced